checkpoint: renames "Base64/exports.ts" to "Base64/Conformance/exports.ts"
