+++ src/library/Sequence/Base64/index.ts
@@ -1,1 +1,1 @@
-export * as Sequence_Base64 from './exports.ts';
+export * as Sequence_Base64 from './Conformance/exports.ts';
